(PC-12795)[PRO] feat: move to sib - booking cancellation email by user to pro
